+++ portal-impl/src/com/liferay/portal/verify/VerifyProcessSuite.java
@@ -32,7 +32,6 @@ public class VerifyProcessSuite extends VerifyProcess {
 		verify(new VerifyResourceActions());
 		verify(new VerifyResourcePermissions());
 		verify(new VerifyUser());
-		verify(new VerifyWorkflow());
 	}
 
 }
